+++ lab_1/key1.json
@@ -1,0 +1,34 @@
+{ 
+    "Ё": "Ё",
+    "Ж": "Ж",
+    "П": "П",
+    "Р": "Р",
+    "=": "Ъ",
+    "<": "Ы",
+    "N": "М",
+    "Z": "Н",
+    "Y": "С",
+    "7": "А",
+    "1": "Б",
+    "9": "В",
+    "3": "Е",
+    "Q": "З",
+    "Я": "Я",
+    "Д": "Д",
+    "г": "Г",
+    "у": "У",
+    "О": "О",
+    "И": "И",
+    "Й": "Й",
+    "Х": "Х",
+    "s": "Ф",
+    "R": "Ц",
+    "%": "Ь",
+    "Т": "Т",
+    "К": "К",
+    "U": "Л",
+    "i": "Ч",
+    "F": "Ш",
+    "G": "Щ",
+    "ю": "Ю"
+}
