+++ lab_1/key2.json
@@ -1,0 +1,34 @@
+{
+    "Ц": "ю",
+    "Й": "б",
+    "Л": "т",
+    "М": "ь",
+    "П": "с",
+    "Ы": "ч",
+    "З": "е",
+    "У": "д",
+    "Ю": "х",
+    "Щ": "э",
+    "Г": "ц",
+    "А": "о",
+    "Х": "м",
+    "Р": "к",
+    "И": "п",
+    "Б": "в",
+    "Ж": "н",
+    "Н": "ф",
+    "Ш": "щ",
+    "Ё": "и",
+    "Я": "а",
+    "Д": "я",
+    "В": "з",
+    "О": " ",
+    "Ь": "г",
+    "Ч": "ш",
+    "Т": "л",
+    "К": "ы",
+    "Е": "й",
+    "Ф": "у",
+    "Ъ": "р",
+    "С": "ж"
+}
