+++ lab_1/task1key.json
@@ -30,5 +30,5 @@
     "Ь": "Б",
     "Э": "В",
     "Ю": "Г",
-    "Я": "Д",
+    "Я": "Д"
 }
