+++ lab_1/task2key.json
@@ -5,7 +5,7 @@
 	"Z": "ф", 
 	"К": "ы", 
 	"N": "ь", 
-	"О": " "
+	"О": " ",
 	"F": "щ", 
 	"Ё": "и",
 	"у": "д", 
@@ -30,5 +30,5 @@
 	"Х": "м", 
 	"Р": "к", 
 	"Д": "я", 
-	"Т": "л", 
+	"Т": "л"
 }
